requirements.txt
pandas
numpy
nltk
vaderSentiment
matplotlib
wordcloud
streamlit
pytest
python-dateutil
Pillow
transformers>=4.30.0
torch>=2.0.0
sentencepiece
accelerate
langdetect
emoji>=2.0.0
plotly
nrclex
